homework6/services/order/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
psycopg2
drf-yasg